create voice service for music commands
